Make server members menu its own component
